scrapy_demo/requirements.txt
aniso8601==1.2.0
APScheduler==3.3.1
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.8.0
cffi==1.12.3
click==6.7
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
Django==2.2.4
djangorestframework==3.10.2
Flask==0.12.1
Flask-BasicAuth==0.2.0
Flask-RESTful==0.3.5
flask-restful-swagger==0.19
Flask-SQLAlchemy==2.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.9.6
lxml==4.4.0
MarkupSafe==1.0
#mysqlclient==1.4.2.post1
parsel==1.5.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyMySQL==0.7.11
pyOpenSSL==19.0.0
python-dateutil==2.6.0
pytz==2017.2
queuelib==1.5.0
requests==2.13.0
Scrapy==1.7.3
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.10.0
soupsieve==1.9.2
SpiderKeeper==1.2.0
SQLAlchemy==1.1.9
sqlparse==0.3.0
Twisted==19.7.0
tzlocal==1.3
w3lib==1.20.0
Werkzeug==0.12.1
zope.interface==4.6.0
